requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pytest==8.1.1
jinja2==3.1.3
python-multipart==0.0.9
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
-e .
